backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-learn==1.4.2
pillow==10.3.0
sqlalchemy==2.0.30

# Training dependencies
tensorflow==2.15.0
librosa==0.10.1
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
